duckuments/requirements.txt
django
djangorestframework
Pillow
djangorestframework-simplejwt
openai
django-cors-headers